Adjusting landing page upsell setting to only load if using Advanced Levels Page Shortcode Add On

+++ elements/page_banners.php
@@ -52,7 +52,7 @@ function memberlite_elements_settings_meta_box_callback( $post ) {
 	$memberlite_landing_page_checkout_button = get_post_meta($post->ID, '_memberlite_landing_page_checkout_button', true);
 	$pmproal_landing_page_level = get_post_meta($post->ID, '_pmproal_landing_page_level', true);
 	$memberlite_landing_page_upsell = get_post_meta($post->ID, '_memberlite_landing_page_upsell', true);
-	echo '<h2><strong>' . __('Page Banner Settings', 'memberlite-elements') . '</strong></h2>';
+	echo '<h3>' . __('Page Banner Settings', 'memberlite-elements') . '</h3>';
 	echo '<p style="margin: 1rem 0 0 0;"><strong>' . __('Show Page Banner', 'memberlite-elements') . '</strong> <em>Disable the entire page banner for this content.</em><br />';
 	echo '<label class="screen-reader-text" for="memberlite_banner_show">';
 	_e('Show Page Banner', 'memberlite-elements');
@@ -107,7 +107,7 @@ function memberlite_elements_settings_meta_box_callback( $post ) {
 	if(($memberlite_page_template == 'templates/landing.php') && function_exists('pmpro_getAllLevels'))
 	{
 		echo '<hr />';
-		echo '<h2>' . __('Landing Page Settings', 'memberlite-elements') . '</h2>';
+		echo '<h3>' . __('Landing Page Settings', 'memberlite-elements') . '</h3>';
 		$membership_levels = pmpro_getAllLevels();
 		if(empty($membership_levels))
 			echo '<div class="inline notice error"><p><a href="' . admin_url('admin.php?page=pmpro-membershiplevels') . '">Add a Membership Level to Use These Landing Page Features &raquo;</a></p>';
@@ -131,16 +131,28 @@ function memberlite_elements_settings_meta_box_callback( $post ) {
 			echo '</label>';
 			echo '<input type="text" id="memberlite_landing_page_checkout_button" name="memberlite_landing_page_checkout_button" value="' . $memberlite_landing_page_checkout_button . '"> <em>(default: "Select")</em></td></tr>';
 			echo '<tr><th scope="row">' . __('Membership Level Upsell', 'memberlite-elements') . '</th>';
-			echo '<td><label class="screen-reader-text" for="memberlite_landing_page_upsell">';
-				_e('Landing Page Membership Level Upsell', 'memberlite-elements');
-			echo '</label>';
-			echo '<select id="memberlite_landing_page_upsell" name="memberlite_landing_page_upsell">';
-			echo '<option value="" ' . selected( $memberlite_landing_page_upsell, "" ) . '>- Select -</option>';
-			foreach($membership_levels as $level)
-			{
-				echo '<option value="' . $level->id . '"' . selected( $memberlite_landing_page_upsell, $level->id ) . '>' . $level->name . '</option>';
+			if ( ! function_exists( 'pmpro_advanced_levels_shortcode' ) ) {
+				$allowed_advanced_levels_html = array (
+					'a' => array (
+						'href' => array(),
+						'target' => array(),
+						'title' => array(),
+					),
+				);
+				echo '<td><p class="description">' . sprintf( wp_kses( __( 'Optional: You must install and activate the <a href="%s" title="Paid Memberships Pro - Advanced Levels Page Add On" target="_blank">Advanced Levels Page Add On</a> to use this landing page feature.', 'memberlite-elements' ), $allowed_advanced_levels_html ), 'https://www.paidmembershipspro.com/add-ons/pmpro-advanced-levels-shortcode/' ) . '</td>';
+			} else {
+				echo '<td><label class="screen-reader-text" for="memberlite_landing_page_upsell">';
+					_e('Landing Page Membership Level Upsell', 'memberlite-elements');
+				echo '</label>';
+				echo '<select id="memberlite_landing_page_upsell" name="memberlite_landing_page_upsell">';
+				echo '<option value="" ' . selected( $memberlite_landing_page_upsell, "" ) . '>- Select -</option>';
+				foreach($membership_levels as $level)
+				{
+					echo '<option value="' . $level->id . '"' . selected( $memberlite_landing_page_upsell, $level->id ) . '>' . $level->name . '</option>';
+				}
+				echo '</select></td>';
 			}
-			echo '</select></td></tr>';
+			echo '</tr>';
 			echo '</tbody></table>';
 		}
 	}
